// miniprogram/api/generated/models/NotificationItemDto.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NotificationItemDto
 */
export interface NotificationItemDto {
    /**
     * 
     * @type {number}
     * @memberof NotificationItemDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof NotificationItemDto
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationItemDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationItemDto
     */
    content?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationItemDto
     */
    refType?: string;
    /**
     * 
     * @type {number}
     * @memberof NotificationItemDto
     */
    refId?: number;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationItemDto
     */
    isRead?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof NotificationItemDto
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the NotificationItemDto interface.
 */
export function instanceOfNotificationItemDto(value: object): value is NotificationItemDto {
    return true;
}

export function NotificationItemDtoFromJSON(json: any): NotificationItemDto {
    return NotificationItemDtoFromJSONTyped(json, false);
}

export function NotificationItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'type': json['type'] == null ? undefined : json['type'],
        'title': json['title'] == null ? undefined : json['title'],
        'content': json['content'] == null ? undefined : json['content'],
        'refType': json['ref_type'] == null ? undefined : json['ref_type'],
        'refId': json['ref_id'] == null ? undefined : json['ref_id'],
        'isRead': json['is_read'] == null ? undefined : json['is_read'],
        'createdAt': json['created_at'] == null ? undefined : (new Date(json['created_at'])),
    };
}

export function NotificationItemDtoToJSON(json: any): NotificationItemDto {
    return NotificationItemDtoToJSONTyped(json, false);
}

export function NotificationItemDtoToJSONTyped(value?: NotificationItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'type': value['type'],
        'title': value['title'],
        'content': value['content'],
        'ref_type': value['refType'],
        'ref_id': value['refId'],
        'is_read': value['isRead'],
        'created_at': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
    };
}